Moved identifier based config to regular config package along with tests from a separate file to the main config test file. Removed all remnants of the old config system
