--- medcat/idconfig.py
@@ -1,514 +1,0 @@
-from datetime import datetime
-from enum import Enum
-from pydantic import BaseModel, Field, Extra
-from pydantic.dataclasses import Any, Callable, Dict, Optional, Union
-from typing import List
-from multiprocessing import cpu_count
-import logging
-import jsonpickle
-from functools import partial
-import re
-
-from medcat.utils.hasher import Hasher
-
-LOGGER = logging.getLogger(__package__)
-
-
-def weighted_average(step: int, factor: float) -> float:
-    return max(0.1, 1 - (step ** 2 * factor))
-
-
-def workers(workers_override: Optional[int] = None) -> int:
-    return max(cpu_count() - 1, 1) if workers_override is None else workers_override
-
-
-class FakeDict(object):
-    """FakeDict that allows the use of the __get__ method for old access tokens.
-    """
-
-    def __getitem__(self, arg: str) -> Any:
-        return getattr(self, arg)
-
-    def __setitem__(self, arg: str, val):
-        return setattr(self, arg, val)
-
-    def get(self, key, default=None) -> Any:
-        try:
-            return self[key]
-        except KeyError:
-            return default
-
-
-class MixingConfig(FakeDict):
-    """Config that is able to saved and loaded from disk as well as mixed with other configs.
-    It is not intended to be initialised directly and it is assumed that instances are dataclasses
-    """
-
-    def save(self, save_path: str) -> None:
-        r''' Save the config into a .json file
-
-        Args:
-            save_path (`str`):
-                Where to save the created json file
-        '''
-        # We want to save the dict here, not the whole class
-        json_string = jsonpickle.encode(
-            {field: getattr(self, field) for field in self.fields()})
-
-        with open(save_path, 'w') as f:
-            f.write(json_string)
-
-    def merge_config(self, config_dict: Dict) -> None:
-        r''' Merge a config_dict with the existing config object.
-
-        Args:
-            config_dict (`dict`):
-                A dictionary which key/values should be added to this class.
-        '''
-        if not isinstance(config_dict, dict): # TODO - remove
-            raise ValueError('Need DICT (for now!)')
-        for key in config_dict.keys():
-            if hasattr(self, key):
-                attr = getattr(self, key)
-            else:
-                attr = None # new attribute
-            value = config_dict[key]
-            if isinstance(value, MixingConfig):
-                value = value.dict()
-            if isinstance(attr, MixingConfig):
-                attr.merge_config(value)
-            else:
-                if isinstance(self, BaseModel):
-                    def is_correct_type(field_type, obj):
-                        obj_type = type(obj)
-                        if field_type is obj_type or field_type == obj_type:
-                            return True
-                        if 'Optional' in str(field_type) and obj is None:
-                            return True
-                        if field_type is Any:
-                            return True
-                        if field_type is Callable and hasattr(obj, '__call__'):
-                            return True
-                        return False
-                    all_fields = self.fields()
-                    found = [f for f in all_fields if f == key]
-                    if len(found) > 0:
-                        found = all_fields[found[0]]
-                        if not is_correct_type(found.type_, value):
-                            LOGGER.info('DEBUG: attr is NOT MixingConfig (but am within BaseModel) - SETTING (BAD)', key, 'expected', found.type_, 'setting', type(value), 'w', found)
-                    else:
-                        LOGGER.info('DEBUG: Could not find BaseModel field for', key, 'in', type(self))#, 'SKIPPING')
-                        # continue
-                try:
-                    setattr(self, key, value)
-                except AttributeError as e:
-                    LOGGER.warning('DEBUG: Issue with setting attribtue', key, ':', e)
-        self.rebuild_re()
-
-    def parse_config_file(self, path: str) -> None:
-        r'''
-        Parses a configuration file in text format. Must be like:
-                cat.<variable>.<key> = <value>
-                ...
-
-            - variable: linking, general, ner, ...
-            - key: a key in the config dict e.g. subsample_after for linking
-            - value: the value for the key, will be parsed with `eval`
-        '''
-        with open(path, 'r') as f:
-            for line in f:
-                if line.strip() and line.startswith("cat."):
-                    line = line[4:]
-                    left, right = line.split("=")
-                    variable, key = left.split(".")
-                    variable = variable.strip()
-                    key = key.strip()
-                    value = eval(right.strip())
-
-                    attr = getattr(self, variable)
-                    if isinstance(attr, MixingConfig):
-                        setattr(attr, key, value)  # assuming
-                    elif isinstance(attr, dict):
-                        attr[key] = value
-                    else:
-                        raise ValueError(
-                            f'Unknown attribute {attr} for "{line}"')
-
-        self.rebuild_re()
-
-    def rebuild_re(self) -> None:
-        pass
-
-    def _calc_hash(self, hasher: Optional[Hasher] = None) -> Hasher:
-        if hasher is None:
-            hasher = Hasher()
-        for _, v in self.dict().items():
-            if isinstance(v, MixingConfig):
-                v._calc_hash(hasher)
-            else:
-                hasher.update(v)
-        return hasher
-
-    def get_hash(self, hasher: Optional[Hasher] = None):
-        hasher = self._calc_hash(hasher)
-        return hasher.hexdigest()
-
-    def __str__(self) -> str:
-        return str(self.dict())
-
-    @classmethod
-    def load(cls, save_path: str) -> "MixingConfig":
-        r''' Load config from a json file, note that fields that
-        did not exist in the old config but do exist in the current
-        version of the ConfigMetaCAT class will be kept.
-
-        Args:
-            save_path (`str`):
-                Path to the json file to load
-        '''
-        config = cls()
-
-        # Read the jsonpickle string
-        with open(save_path) as f:
-            config_dict = jsonpickle.decode(f.read())
-
-        config.merge_config(config_dict)
-
-        return config
-
-    @classmethod
-    def from_dict(cls, config_dict: Dict) -> "MixingConfig":
-        config = cls()
-        config.merge_config(config_dict)
-        return config
-
-    def asdict(self):
-        return self.dict()
-
-    def keys(self) -> List[str]:
-        return self.fields().keys()
-
-    def fields(self) -> Dict[str, Field]:
-        return self.__fields__
-
-
-class VersionInfo(MixingConfig, BaseModel):
-    # rearranging order so that "non-default" values (with field method) 
-    # appear before ones with explicit default values
-    history: list = Field(default_factory=list)  # Populated automatically
-    meta_cats: dict = Field(default_factory=dict)  # Populated automatically
-    # Populated automatically, output from cdb.print_stats
-    cdb_info: dict = Field(default_factory=dict)
-    # NER general performance, meta should be: {'meta': {'model_name': {'f1': <>, 'p': <>, ...}, ...}}
-    performance: dict = Field(default_factory=lambda: {'ner': {}, 'meta': {}})
-    # General description and what it was trained on
-    description: str = "No description"
-    id: Any = None  # Will be: hash of most things
-    last_modified: Optional[Union[int, datetime]] = None  # Yep
-    # Path/URL/Whatever to where is this CDB located
-    location: Optional[str] = None
-    # What was used to build the CDB, e.g. SNOMED_202009
-    ontology: Optional[str] = None
-    # Which version of medcat was used to build the CDB
-    medcat_version: Optional[str] = None
-
-    class Config:
-        extra = Extra.allow
-
-
-class NameVersion(Enum):
-    LOWER = 1
-    CLEAN = 2
-
-
-class CDBMaker(MixingConfig, BaseModel):
-    # rearranging order so that "non-default" values (with field method) 
-    # appear before ones with explicit default values
-    # Name versions to be generated.
-    name_versions: list = Field(default_factory=lambda: ['LOWER', 'CLEAN'])
-    # If multiple names or type_ids for a concept present in one row of a CSV, they are separted
-    # by the character below.
-    multi_separator: str = '|'
-    # Should preferred names with parenthesis be cleaned 0 means no, else it means if longer than or equal
-    # e.g. Head (Body part) -> Head
-    remove_parenthesis: int = 5
-    # Minimum number of letters required in a name to be accepted for a concept
-    min_letters_required: int = 2
-
-    class Config:
-        extra = Extra.allow
-
-
-class AnnotationOutput(MixingConfig, BaseModel):
-    doc_extended_info: bool = False
-    context_left: int = -1
-    context_right: int = -1
-    lowercase_context: bool = True
-    include_text_in_output: bool = False
-
-    class Config:
-        extra = Extra.allow
-
-
-class LabelType(str, Enum):
-    short = 1
-    long = 2
-
-
-class CheckPoint(MixingConfig, BaseModel):
-    output_dir: str = 'checkpoints'
-    """When doing training this is the name of the directory where checkpoints will be saved"""
-    steps: Optional[int] = None
-    """When training how often to save the checkpoint (one step represents one document), if None no ckpts will be created"""
-    max_to_keep: int = 1
-    """When training the maximum checkpoints will be kept on the disk"""
-
-    class Config:
-        extra = Extra.allow
-
-
-class General(MixingConfig, BaseModel):
-    # rearranging order so that "non-default" values (with field method) 
-    # appear before ones with explicit default values
-    spacy_disabled_components: list = Field(default_factory=lambda: ['ner', 'parser', 'vectors', 'textcat',
-                                                                     'entity_linker', 'sentencizer', 'entity_ruler', 'merge_noun_chunks',
-                                                                     'merge_entities', 'merge_subtokens'])
-    checkpoint: CheckPoint = CheckPoint()
-    """Checkpointing config"""
-    log_level: int = logging.INFO
-    """Logging config for everything | 'tagger' can be disabled, but will cause a drop in performance"""
-    log_format: str = '%(levelname)s:%(name)s: %(message)s'
-    log_path: str = './medcat.log'
-    spacy_model: str = 'en_core_web_md'
-    """What model will be used for tokenization"""
-    separator: str = '~'
-    """Separator that will be used to merge tokens of a name. Once a CDB is built this should
-    always stay the same."""
-    spell_check: bool = True
-    """Should we check spelling - note that this makes things much slower, use only if necessary. The only thing necessary
-    for the spell checker to work is vocab.dat and cdb.dat built with concepts in the respective language."""
-    diacritics: bool = False
-    """Should we process diacritics - for languages other than English, symbols such as 'é, ë, ö' can be relevant.
-    Note that this makes spell_check slower."""
-    spell_check_deep: bool = False
-    """If True the spell checker will try harder to find mistakes, this can slow down
-    things drastically."""
-    spell_check_len_limit: int = 7
-    """Spelling will not be checked for words with length less than this"""
-    show_nested_entities: bool = False
-    """If set to True functions like get_entities and get_json will return nested_entities and overlaps"""
-    full_unlink: bool = False
-    """When unlinking a name from a concept should we do full_unlink (means unlink a name from all concepts, not just the one in question)"""
-    workers: int = workers()
-    """Number of workers used by a parallelizable pipeline component"""
-    make_pretty_labels: Optional[str] = None # TODO - consider moving to LabelType enum
-    """Should the labels of entities (shown in displacy) be pretty or just 'concept'. Slows down the annotation pipeline
-    should not be used when annotating millions of documents. If `None` it will be the string "concept", if `short` it will be CUI,
-    if `long` it will be CUI | Name | Confidence"""
-    # make_pretty_labels: Optional[LabelType] = None
-    map_cui_to_group: bool = False
-    """If the cdb.addl_info['cui2group'] is provided and this option enabled, each CUI will be maped to the group"""
-
-    class Config:
-        extra = Extra.allow
-
-
-class Preprocessing(MixingConfig, BaseModel):
-    # rearranging order so that "non-default" values (with field method) 
-    # appear before ones with explicit default values
-    words_to_skip: set = Field(default_factory=lambda: {'nos'})
-    """This words will be completly ignored from concepts and from the text (must be a Set)"""
-    keep_punct: set = Field(default_factory=lambda: {'.', ':'})
-    """All punct will be skipped by default, here you can set what will be kept"""
-    do_not_normalize: set = Field(default_factory=lambda: {
-                                  'VBD', 'VBG', 'VBN', 'VBP', 'JJS', 'JJR'})
-    """Should specific word types be normalized: e.g. running -> run
-    Values are detailed part-of-speech tags. See:
-    - https://spacy.io/usage/linguistic-features#pos-tagging
-    - Label scheme section per model at https://spacy.io/models/en"""
-    skip_stopwords: bool = False
-    """Should stopwords be skipped/ingored when processing input"""
-    min_len_normalize: int = 5
-    """Nothing below this length will ever be normalized (input tokens or concept names), normalized means lemmatized in this case"""
-    stopwords: Optional[set] = None
-    """If None the default set of stowords from spacy will be used. This must be a Set."""
-    max_document_length: int = 1000000
-    """Documents longer  than this will be trimmed"""
-
-    class Config:
-        extra = Extra.allow
-
-
-class Ner(MixingConfig, BaseModel):
-    min_name_len: int = 3
-    """Do not detect names below this limit, skip them"""
-    max_skip_tokens: int = 2
-    """When checkng tokens for concepts you can have skipped tokens inbetween
-    used ones (usually spaces, new lines etc). This number tells you how many skipped can you have."""
-    check_upper_case_names: bool = False
-    """Check uppercase to distinguish uppercase and lowercase words that have a different meaning."""
-    upper_case_limit_len: int = 4
-    """Any name shorter than this must be uppercase in the text to be considered. If it is not uppercase
-    it will be skipped."""
-    try_reverse_word_order: bool = False
-    """Try reverse word order for short concepts (2 words max), e.g. heart disease -> disease heart"""
-
-    class Config:
-        extra = Extra.allow
-
-
-class _DefPartial(object):
-    """This is a helper class to make it possible to check equality of two diffault Linking instances
-    """
-
-    def __init__(self):
-        self.fun = partial(weighted_average, factor=0.0004)
-
-    def __call__(self, *args, **kwargs):
-        return self.fun(*args, **kwargs)
-
-    def __eq__(self, other):
-        return isinstance(other, _DefPartial)
-
-
-_DEFAULT_PARTIAL = _DefPartial()
-
-
-class Linking(MixingConfig, BaseModel):
-    # rearranging order so that "non-default" values (with field method) 
-    # appear before ones with explicit default values
-    optim: dict = Field(default_factory=lambda: {
-                        'type': 'linear', 'base_lr': 1, 'min_lr': 0.00005})
-    """Linear anneal"""
-    # optim: dict = {'type': 'standard', 'lr': 1}
-    # optim: dict = {'type': 'moving_avg', 'alpha': 0.99, 'e': 1e-4, 'size': 100}
-    context_vector_sizes: dict = Field(
-        default_factory=lambda: {'xlong': 27, 'long': 18, 'medium': 9, 'short': 3})
-    """Context vector sizes that will be calculated and used for linking"""
-    context_vector_weights: dict = Field(default_factory=lambda: {
-                                         'xlong': 0.1, 'long': 0.4, 'medium': 0.4, 'short': 0.1})
-    """Weight of each vector in the similarity score - make trainable at some point. Should add up to 1."""
-    filters: dict = Field(default_factory=lambda: {
-        'cuis': set(),  # CUIs in this filter will be included, everything else excluded, must be a set, if empty all cuis will be included
-    })
-    """Filters"""
-    train: bool = True
-    """Should it train or not, this is set automatically ignore in 99% of cases and do not set manually"""
-    random_replacement_unsupervised: float = 0.80
-    """If <1 during unsupervised training the detected term will be randomly replaced with a probability of 1 - random_replacement_unsupervised
-    Replaced with a synonym used for that term"""
-    disamb_length_limit: int = 3
-    """All concepts below this will always be disambiguated"""
-    filter_before_disamb: bool = False
-    """If True it will filter before doing disamb. Useful for the trainer."""
-    train_count_threshold: int = 1
-    """Concepts that have seen less training examples than this will not be used for
-    similarity calculation and will have a similarity of -1."""
-    always_calculate_similarity: bool = False
-    """Do we want to calculate context similarity even for concepts that are not ambigous."""
-    weighted_average_function: Callable = _DEFAULT_PARTIAL
-    """Weights for a weighted average
-    'weighted_average_function': partial(weighted_average, factor=0.02),"""
-    calculate_dynamic_threshold: bool = False
-    """Concepts below this similarity will be ignored. Type can be static/dynamic - if dynamic each CUI has a different TH
-    and it is calcualted as the average confidence for that CUI * similarity_threshold. Take care that dynamic works only
-    if the cdb was trained with calculate_dynamic_threshold = True."""
-    similarity_threshold_type: str = 'static'
-    similarity_threshold: float = 0.25
-    negative_probability: float = 0.5
-    """Probability for the negative context to be added for each positive addition"""
-    negative_ignore_punct_and_num: bool = True
-    """Do we ignore punct/num when negative sampling"""
-    prefer_primary_name: float = 0.35
-    """If >0 concepts for which a detection is its primary name will be preferred by that amount (0 to 1)"""
-    prefer_frequent_concepts: float = 0.35
-    """If >0 concepts that are more frequent will be prefered by a multiply of this amount"""
-    subsample_after: int = 30000
-    """DISABLED in code permanetly: Subsample during unsupervised training if a concept has received more than"""
-    devalue_linked_concepts: bool = False
-    """When adding a positive example, should it also be treated as Negative for concepts
-    which link to the postive one via names (ambigous names)."""
-    context_ignore_center_tokens: bool = False
-    """If true when the context of a concept is calculated (embedding) the words making that concept are not taken into accout"""
-
-    class Config:
-        extra = Extra.allow
-
-
-class BaseConfig(MixingConfig, BaseModel):
-    version: VersionInfo = VersionInfo()
-    cdb_maker: CDBMaker = CDBMaker()
-    annotation_output: AnnotationOutput = AnnotationOutput()
-    general: General = General()
-    preprocessing: Preprocessing = Preprocessing()
-    ner: Ner = Ner()
-    linking: Linking = Linking()
-    word_skipper: Optional[re.Pattern] = None
-    punct_checker: Optional[re.Pattern] = None
-
-    class Config:
-        # this if for word_skipper and punct_checker which would otherwise
-        # not have a validator
-        arbitrary_types_allowed = True
-        extra = Extra.allow
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.rebuild_re()
-
-    # Override
-    def rebuild_re(self) -> None:
-        # Some regex that we will need
-        self.word_skipper = re.compile('^({})$'.format(
-            '|'.join(self.preprocessing.words_to_skip)))
-        # Very agressive punct checker, input will be lowercased
-        self.punct_checker = re.compile(r'[^a-z0-9]+')
-
-    # Override
-    def get_hash(self):
-        hasher = Hasher()
-        for k, v in self.dict().items():
-            if k not in ['version', 'general', 'linking']:
-                hasher.update(v, length=True)
-            elif k == 'general':
-                for k2, v2 in v.items():
-                    if k2 != 'spacy_model':
-                        hasher.update(v2, length=False)
-                    else:
-                        # Ignore spacy model
-                        pass
-            elif k == 'linking':
-                for k2, v2 in v.items():
-                    if k2 != "filters":
-                        hasher.update(v2, length=False)
-                    else:
-                        hasher.update(v2, length=True)
-
-        return hasher.hexdigest()
-
-
-if __name__ == '__main__':
-    print('START')
-    bc: BaseConfig = BaseConfig()
-    print('BaseConfig', bc)
-    # print('as dict', asdict(bc))
-    print('as dict', bc.dict())
-    print('with re', bc.word_skipper, 'and', bc.punct_checker)
-    print('\n')
-    print('Attempting to get with ["key"]:', bc.general['log_level'])
-    print('\n')
-    print('Now attempting to SET with ["key"]')
-    bc.general['log_level'] = logging.WARN
-    print('Set to', bc.general.log_level)
-    print('\n')
-    print('CORRECT setting')
-    bc.general.log_level = logging.INFO
-    print('Now at', bc.general.log_level)
-
-    print('Testing MERGING')
-    g1: General = General(spacy_model='model_1')
-    g2: General = General(spacy_model='/tmp/tm123/model_2')
-    print(g1.spacy_model, 'vs', g2.spacy_model)
-    # g1.merge_config(asdict(g2))
-    g1.merge_config(g2.dict())
-    print(g1.spacy_model, 'vs', g2.spacy_model)
